Fixing write-back behavior for mutable structs in lambda compiler.
