+++ src/System.Linq.Expressions/src/System/Linq/Expressions/Compiler/LambdaCompiler.Address.cs
@@ -277,9 +277,10 @@ namespace System.Linq.Expressions.Compiler
             if (node.Expression != null)
             {
                 EmitInstance(node.Expression, instanceType = node.Expression.Type);
+
                 // store in local
                 _ilg.Emit(OpCodes.Dup);
-                _ilg.Emit(OpCodes.Stloc, instanceLocal = GetLocal(instanceType));
+                _ilg.Emit(OpCodes.Stloc, instanceLocal = GetInstanceLocal(instanceType));
             }
 
             PropertyInfo pi = (PropertyInfo)node.Member;
@@ -321,8 +322,9 @@ namespace System.Linq.Expressions.Compiler
             {
                 EmitInstance(node.Object, instanceType = node.Object.Type);
 
+                // store in local
                 _ilg.Emit(OpCodes.Dup);
-                _ilg.Emit(OpCodes.Stloc, instanceLocal = GetLocal(instanceType));
+                _ilg.Emit(OpCodes.Stloc, instanceLocal = GetInstanceLocal(instanceType));
             }
 
             // Emit indexes. We don't allow byref args, so no need to worry
@@ -365,6 +367,12 @@ namespace System.Linq.Expressions.Compiler
 
                 EmitSetIndexCall(node, instanceType);
             };
+        }
+
+        private LocalBuilder GetInstanceLocal(Type type)
+        {
+            var instanceLocalType = type.GetTypeInfo().IsValueType ? type.MakeByRefType() : type;
+            return GetLocal(instanceLocalType);
         }
     }
 }
