+++ src/System.Linq.Expressions/tests/System.Linq.Expressions.Tests.csproj
@@ -85,6 +85,7 @@
     <Compile Include="Block\ParameterBlockTests.cs" />
     <Compile Include="Block\SharedBlockTests.cs" />
     <Compile Include="Call\CallFactoryTests.cs" />
+    <Compile Include="Call\CallTests.cs" />
     <Compile Include="Cast\AsNullable.cs" />
     <Compile Include="Cast\AsTests.cs" />
     <Compile Include="Cast\CastNullableTests.cs" />
